+++ src/meta_extra.zig
@@ -13,10 +13,10 @@ pub inline fn isMarked(comptime T: type, comptime name: []const u8) bool {
     return false;
 }
 
-pub inline fn checkIterContraints(comptime T: type, comptime name: []const u8, comptime Item: type) bool {
-    if (!@hasField(T, "wrapped")) return false;
-    const Wrapped = @FieldType(T, "wrapped");
-    return comptime T == Iter(Wrapped) and Wrapped.Item == Item and isMarked(Wrapped, name);
+pub inline fn checkIterConstraints(comptime I: type, comptime name: []const u8, comptime Item: type) bool {
+    if (!@hasField(I, "wrapped")) return false;
+    const Wrapped = @FieldType(I, "wrapped");
+    return comptime I == Iter(Wrapped) and Wrapped.Item == Item and isMarked(Wrapped, name);
 }
 
 pub inline fn expectImplIter(comptime T: type) !void {