+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Export `checkIterConstraints()`
+
 ### Changed
 
 - Pass `self` directly to `Cycle.wrapped.iter` in `Iter.cycle`